Illuminating:I switched back to the initial SFINAE (see comment)
From: https://en.cppreference.com/w/cpp/language/sfinae

Substitution proceeds in lexical order and stops when a failure is encountered.

================================================================
template <typename A>
struct B { using type = typename A::type; };

template <
  class T,
  class U = typename T::type,      // SFINAE failure if T has no member type
  class V = typename B<T>::type    // hard error if T has no member type
                                   // (guaranteed to not occur as of C++14 because
                                   //  substitution into the default template argument
                                   //  of U would fail first)
> void foo (int);
================================================================

If there are multiple declarations with different lexical
orders (e.g. a function template declared with trailing return type,
to be substituted after a parameter, and redeclared with ordinary
return type that would be substituted before the parameter), and that
would cause template instantiations to occur in a different order or
not at all, then the program is ill-formed; no diagnostic required.